serverless/requirements.txt
requests==2.31.0
torch==2.2.1
torchvision==0.17.1
streamlit==1.32.2 # 1.32.2
google_api_python_client==1.12.8
protobuf==3.20.3
opencv-python==4.9.0.80
pillow==10.2.0